shape ops work in progress
git-svn-id: http://skia.googlecode.com/svn/trunk@4119 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ experimental/Intersection/QuadraticIntersection_Test.cpp
@@ -54,8 +54,8 @@ static const Quadratic testSet[] = {
 const size_t testSetCount = sizeof(testSet) / sizeof(testSet[0]);
 
 static void oneOffTest() {
-    for (int outer = 0; outer < testSetCount - 1; ++outer) {
-        for (int inner = outer + 1; inner < testSetCount; ++inner) {
+    for (size_t outer = 0; outer < testSetCount - 1; ++outer) {
+        for (size_t inner = outer + 1; inner < testSetCount; ++inner) {
             const Quadratic& quad1 = testSet[outer];
             const Quadratic& quad2 = testSet[inner];
             Intersections intersections;
